+++ lib/MailMergeService.php
@@ -1,0 +1,82 @@
+<?php
+/**
+ *
+ * (c) Copyright Ascensio System SIA 2026
+ *
+ * This program is a free software product.
+ * You can redistribute it and/or modify it under the terms of the GNU Affero General Public License
+ * (AGPL) version 3 as published by the Free Software Foundation.
+ * In accordance with Section 7(a) of the GNU AGPL its Section 15 shall be amended to the effect
+ * that Ascensio System SIA expressly excludes the warranty of non-infringement of any third-party rights.
+ *
+ * This program is distributed WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * For details, see the GNU AGPL at: http://www.gnu.org/licenses/agpl-3.0.html
+ *
+ * You can contact Ascensio System SIA at 20A-12 Ernesta Birznieka-Upisha street, Riga, Latvia, EU, LV-1050.
+ *
+ * The interactive user interfaces in modified source and object code versions of the Program
+ * must display Appropriate Legal Notices, as required under Section 5 of the GNU AGPL version 3.
+ *
+ * Pursuant to Section 7(b) of the License you must retain the original Product logo when distributing the program.
+ * Pursuant to Section 7(e) we decline to grant you any rights under trademark law for use of our trademarks.
+ *
+ * All the Product's GUI elements, including illustrations and icon sets, as well as technical
+ * writing content are licensed under the terms of the Creative Commons Attribution-ShareAlike 4.0 International.
+ * See the License terms at http://creativecommons.org/licenses/by-sa/4.0/legalcode
+ *
+ */
+
+namespace OCA\Onlyoffice;
+
+use Exception;
+use OCP\Mail\Provider\Address;
+use OCP\Mail\Provider\Attachment;
+use OCP\Mail\Provider\IManager;
+
+class MailMergeService {
+
+    /**
+     * @param \OCP\Mail\Provider\IManager $mailManager
+     * @param \Psr\Log\LoggerInterface $logger
+     */
+    public function __construct(
+        private IManager $mailManager,
+    ) {}
+
+    /**
+     * Send MailMerge message
+     * @param string $uid
+     * @param \OCA\Onlyoffice\MailMergeMessage $mailMergeMessage
+     * @throws \Exception
+     * @return void
+     */
+    public function send(string $uid, MailMergeMessage $mailMergeMessage): void {
+        $mailService = $this->mailManager->findServiceByAddress($uid, $mailMergeMessage->getFrom());
+
+        if ($mailService === null || !$mailService->capable('MessageSend')) {
+            throw new Exception("MailMerge error: $uid with {$mailMergeMessage->getFrom()} address cannot send messages");
+        }
+
+        $fromAddress = new Address($mailMergeMessage->getFrom());
+        $toAddress = new Address($mailMergeMessage->getTo());
+
+        $message = $mailService->initiateMessage();
+        $message->setFrom($fromAddress)
+            ->setTo($toAddress)
+            ->setSubject($mailMergeMessage->getSubject())
+            ->setBodyHtml($mailMergeMessage->getBodyHtml())
+            ->setBodyPlain($mailMergeMessage->getBodyPlain());
+
+        if ($mailMergeMessage->hasAttachment()) {
+            $attachment = new Attachment(
+                $mailMergeMessage->getAttachment()->getContent(),
+                $mailMergeMessage->getAttachment()->getName(),
+                $mailMergeMessage->getAttachment()->getExtension()
+            );
+            $message->setAttachments($attachment);
+        }
+
+        $mailService->sendMessage($message);
+    }
+}